Refactor AnonSetFileInput
